Refactoring api
