+++ api/preprocess.py
@@ -1,17 +1,7 @@
-from fastapi import FastAPI, Request
 import pandas as pd
-import uvicorn
-import json
-
-app = FastAPI()
 
 
-@app.post('/predict')
-async def process_json(request: Request):
-    contents = await request.body()
-    data = json.loads(contents)
-
-    # Шаг 2: Создание таблиц
+def preprocess(data) -> list | pd.DataFrame:
     tables = {
         "addresses": [],
         "developers": [],
@@ -22,13 +12,11 @@ async def process_json(request: Request):
         "realty_outside": []
     }
 
-    # Шаг 3: Разделение данных
     for record in data:
         for table_name in tables.keys():
             if table_name in record:
                 tables[table_name].append(record[table_name])
 
-    # Шаг 4: Преобразование таблиц в датафреймы
     addresses_df = pd.DataFrame(tables["addresses"])
     developers_df = pd.DataFrame(tables["developers"])
     offers_df = pd.DataFrame(tables["offers"])
@@ -37,10 +25,8 @@ async def process_json(request: Request):
     realty_inside_df = pd.DataFrame(tables["realty_inside"])
     realty_outside_df = pd.DataFrame(tables["realty_outside"])
 
-    addresses_df['lat'] = addresses_df['coordinates'].apply(
-        lambda x: x['lat'] if isinstance(x, dict) else None)
-    addresses_df['lng'] = addresses_df['coordinates'].apply(
-        lambda x: x['lng'] if isinstance(x, dict) else None)
+    addresses_df['lat'] = addresses_df['coordinates'].apply(lambda x: x['lat'] if isinstance(x, dict) else None)
+    addresses_df['lng'] = addresses_df['coordinates'].apply(lambda x: x['lng'] if isinstance(x, dict) else None)
     addresses_df.drop(columns=['coordinates', 'address'], inplace=True)
 
     offers_df = offers_df.dropna(subset=['photos_count'])
@@ -59,11 +45,9 @@ async def process_json(request: Request):
     realty_outside_df['material_type'] = realty_outside_df['material_type'].replace('none', None)
     realty_details_df.drop(columns=['finish_date'], inplace=True)
 
-    main_df = addresses_df.merge(offers_df, on='cian_id', how='inner').merge(
-        offers_details_df, on='cian_id', how='inner')
+    main_df = addresses_df.merge(offers_df, on='cian_id', how='inner').merge(offers_details_df, on='cian_id', how='inner')
 
-    tables_to_left_join = [developers_df, realty_details_df,
-                           realty_inside_df, realty_outside_df]
+    tables_to_left_join = [developers_df, realty_details_df, realty_inside_df, realty_outside_df]
     for table in tables_to_left_join:
         main_df = main_df.merge(table, on='cian_id', how='left')
 
@@ -77,7 +61,6 @@ async def process_json(request: Request):
 
     main_df['total_rate'] = main_df['total_rate'].fillna(4.180479743602584)
     main_df['review_count'] = main_df['review_count'].fillna(1248.7316089939407)
-    main_df['ceiling_height'] = main_df['ceiling_height'].fillna(2.961070749446198)
 
     mean_proportion_ceiling_height = 0.06049278161032404
     formula = main_df['total_area'] * mean_proportion_ceiling_height
@@ -85,19 +68,15 @@ async def process_json(request: Request):
         main_df['ceiling_height'] > 0, formula)
 
     mean_proportion_living_area = 0.5486437974462534
-    main_df['living_area'] = main_df['living_area'].fillna(
-        main_df['total_area'] * mean_proportion_living_area)
+    main_df['living_area'] = main_df['living_area'].fillna(main_df['total_area'] * mean_proportion_living_area)
 
     mean_proportion_kitchen_area = 0.4577734591774278
-    mask = (main_df['total_area'] - main_df['living_area']
-            ).replace(0, pd.NA) * mean_proportion_kitchen_area
+    mask = (main_df['total_area'] - main_df['living_area']).replace(0, pd.NA) * mean_proportion_kitchen_area
     main_df['kitchen_area'] = main_df['kitchen_area'].fillna(mask)
 
     mean_proportion_rooms_count = 0.06657494706605477
-    main_df['rooms_count'] = main_df['rooms_count'].fillna(
-        main_df['living_area'] * mean_proportion_rooms_count).astype(int)
-    main_df['build_year'] = main_df.apply(lambda row: row['finish_year'] if pd.isna(
-        row['build_year']) else row['build_year'], axis=1)
+    main_df['rooms_count'] = main_df['rooms_count'].fillna(main_df['living_area'] * mean_proportion_rooms_count).astype(int)
+    main_df['build_year'] = main_df.apply(lambda row: row['finish_year'] if pd.isna(row['build_year']) else row['build_year'], axis=1)
 
     main_df = main_df.dropna(
         subset=['travel_time', 'views_count', 'kitchen_area', 'build_year']).copy()
@@ -108,10 +87,8 @@ async def process_json(request: Request):
     main_df['garbage_chute'] = main_df['garbage_chute'].astype(bool).fillna(False)
     main_df['is_emergency'] = main_df['is_emergency'].astype(bool).fillna(False)
     main_df['is_apartment'] = main_df['is_apartment'].astype(bool).fillna(False)
-    main_df['is_mortgage_allowed'] = main_df['is_mortgage_allowed'].astype(
-        bool).fillna(False)
-    main_df['renovation_programm'] = main_df['renovation_programm'].astype(
-        bool).fillna(False)
+    main_df['is_mortgage_allowed'] = main_df['is_mortgage_allowed'].astype(bool).fillna(False)
+    main_df['renovation_programm'] = main_df['renovation_programm'].astype(bool).fillna(False)
 
     main_df['photos_count'] = main_df['photos_count'].astype(int)
     main_df['price'] = main_df['price'].astype(int)
@@ -137,16 +114,10 @@ async def process_json(request: Request):
 
     main_df = main_df.drop(columns=['is_reliable', 'heat_type', 'name', 'gas_type', 'parking_type', 'windows_view',
                            'street', 'agent_name', 'house', 'is_duplicate', 'cian_id', 'lat', 'lng', 'price_changes', 'description'], axis=1)
-    if main_df[['district', 'county', 'sale_type']].isna().any().any():
-        raise Exception
+
+    missing_columns = main_df[['district', 'county', 'sale_type']].isna().any()
+    if missing_columns.any():
+        return missing_columns[missing_columns].index.to_list()
 
     X = main_df.drop(columns=['price'])
-    response = {
-        "X": X.to_json(orient='records', force_ascii=False),
-    }
-    return {"message": 'IZI', "X_data": response}
-
-
-if __name__ == "__main__":
-    uvicorn.run(app, host='localhost', port=8000)
-    
+    return X
